+++ sdk/customer-insights/arm-customerinsights/src/customerInsightsManagementClient.ts
@@ -88,9 +88,6 @@ export class CustomerInsightsManagementClient extends coreClient.ServiceClient {
         ? `${options.userAgentOptions.userAgentPrefix} ${packageDetails}`
         : `${packageDetails}`;
 
-    if (!options.credentialScopes) {
-      options.credentialScopes = ["https://management.azure.com/.default"];
-    }
     const optionsWithDefaults = {
       ...defaults,
       ...options,