+++ sdk/network/arm-network-rest/samples-dev/expressRouteCircuitPeeringsDeleteSample.ts
@@ -1,0 +1,45 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT license.
+
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+import createNetworkManagementClient, {
+  ExpressRouteCircuitPeeringsDeleteParameters,
+  getLongRunningPoller,
+} from "@azure-rest/arm-network";
+import { DefaultAzureCredential } from "@azure/identity";
+import * as dotenv from "dotenv";
+
+dotenv.config();
+
+/**
+ * This sample demonstrates how to Deletes the specified peering from the specified express route circuit.
+ *
+ * @summary Deletes the specified peering from the specified express route circuit.
+ * x-ms-original-file: specification/network/resource-manager/Microsoft.Network/stable/2022-05-01/examples/ExpressRouteCircuitPeeringDelete.json
+ */
+async function deleteExpressRouteCircuitPeerings() {
+  const credential = new DefaultAzureCredential();
+  const client = createNetworkManagementClient(credential);
+  const subscriptionId = "";
+  const resourceGroupName = "rg1";
+  const circuitName = "circuitName";
+  const peeringName = "peeringName";
+  const options: ExpressRouteCircuitPeeringsDeleteParameters = {
+    queryParameters: { "api-version": "2022-05-01" },
+  };
+  const initialResponse = await client
+    .path(
+      "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Network/expressRouteCircuits/{circuitName}/peerings/{peeringName}",
+      subscriptionId,
+      resourceGroupName,
+      circuitName,
+      peeringName
+    )
+    .delete(options);
+  const poller = getLongRunningPoller(client, initialResponse);
+  const result = await poller.pollUntilDone();
+  console.log(result);
+}
+
+deleteExpressRouteCircuitPeerings().catch(console.error);