+++ sdk/network/arm-network-rest/samples/v1-beta/javascript/configurationPolicyGroupsDeleteSample.js
@@ -1,0 +1,41 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT license.
+
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+const createNetworkManagementClient = require("@azure-rest/arm-network").default,
+  { getLongRunningPoller } = require("@azure-rest/arm-network");
+const { DefaultAzureCredential } = require("@azure/identity");
+require("dotenv").config();
+
+/**
+ * This sample demonstrates how to Deletes a ConfigurationPolicyGroup.
+ *
+ * @summary Deletes a ConfigurationPolicyGroup.
+ * x-ms-original-file: specification/network/resource-manager/Microsoft.Network/stable/2022-05-01/examples/ConfigurationPolicyGroupDelete.json
+ */
+async function configurationPolicyGroupDelete() {
+  const credential = new DefaultAzureCredential();
+  const client = createNetworkManagementClient(credential);
+  const subscriptionId = "";
+  const resourceGroupName = "rg1";
+  const vpnServerConfigurationName = "vpnServerConfiguration1";
+  const configurationPolicyGroupName = "policyGroup1";
+  const options = {
+    queryParameters: { "api-version": "2022-05-01" },
+  };
+  const initialResponse = await client
+    .path(
+      "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Network/vpnServerConfigurations/{vpnServerConfigurationName}/configurationPolicyGroups/{configurationPolicyGroupName}",
+      subscriptionId,
+      resourceGroupName,
+      vpnServerConfigurationName,
+      configurationPolicyGroupName
+    )
+    .delete(options);
+  const poller = getLongRunningPoller(client, initialResponse);
+  const result = await poller.pollUntilDone();
+  console.log(result);
+}
+
+configurationPolicyGroupDelete().catch(console.error);