+++ sdk/network/arm-network-rest/samples/v1-beta/javascript/expressRouteCircuitsListRoutesTableSample.js
@@ -1,0 +1,43 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT license.
+
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+const createNetworkManagementClient = require("@azure-rest/arm-network").default,
+  { getLongRunningPoller } = require("@azure-rest/arm-network");
+const { DefaultAzureCredential } = require("@azure/identity");
+require("dotenv").config();
+
+/**
+ * This sample demonstrates how to Gets the currently advertised routes table associated with the express route circuit in a resource group.
+ *
+ * @summary Gets the currently advertised routes table associated with the express route circuit in a resource group.
+ * x-ms-original-file: specification/network/resource-manager/Microsoft.Network/stable/2022-05-01/examples/ExpressRouteCircuitRouteTableList.json
+ */
+async function listRouteTables() {
+  const credential = new DefaultAzureCredential();
+  const client = createNetworkManagementClient(credential);
+  const subscriptionId = "";
+  const resourceGroupName = "rg1";
+  const circuitName = "circuitName";
+  const peeringName = "peeringName";
+  const devicePath = "devicePath";
+  const options = {
+    queryParameters: { "api-version": "2022-05-01" },
+  };
+  const initialResponse = await client
+    .path(
+      "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Network/expressRouteCircuits/{circuitName}/peerings/{peeringName}/routeTables/{devicePath}",
+      subscriptionId,
+      resourceGroupName,
+      circuitName,
+      peeringName,
+      devicePath
+    )
+    .post(options);
+  const poller = getLongRunningPoller(client, initialResponse);
+  const result = await poller.pollUntilDone();
+  console.log(result);
+}
+
+listRouteTables().catch(console.error);