+++ sdk/network/arm-network-rest/samples/v1-beta/javascript/vipSwapCreateSample.js
@@ -1,0 +1,42 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT license.
+
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+const createNetworkManagementClient = require("@azure-rest/arm-network").default,
+  { getLongRunningPoller } = require("@azure-rest/arm-network");
+const { DefaultAzureCredential } = require("@azure/identity");
+require("dotenv").config();
+
+/**
+ * This sample demonstrates how to Performs vip swap operation on swappable cloud services.
+ *
+ * @summary Performs vip swap operation on swappable cloud services.
+ * x-ms-original-file: specification/network/resource-manager/Microsoft.Network/stable/2022-05-01/examples/CloudServiceSwapPut.json
+ */
+async function putVipSwapOperation() {
+  const credential = new DefaultAzureCredential();
+  const client = createNetworkManagementClient(credential);
+  const subscriptionId = "";
+  const groupName = "rg1";
+  const resourceName = "testCloudService";
+  const singletonResource = "swap";
+  const options = {
+    body: { properties: { slotType: "Production" } },
+    queryParameters: { "api-version": "2022-05-01" },
+  };
+  const initialResponse = await client
+    .path(
+      "/subscriptions/{subscriptionId}/resourceGroups/{groupName}/providers/Microsoft.Compute/cloudServices/{resourceName}/providers/Microsoft.Network/cloudServiceSlots/{singletonResource}",
+      subscriptionId,
+      groupName,
+      resourceName,
+      singletonResource
+    )
+    .put(options);
+  const poller = getLongRunningPoller(client, initialResponse);
+  const result = await poller.pollUntilDone();
+  console.log(result);
+}
+
+putVipSwapOperation().catch(console.error);