+++ sdk/network/arm-network-rest/samples/v1-beta/typescript/src/customIPPrefixesUpdateTagsSample.ts
@@ -1,0 +1,41 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT license.
+
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+import createNetworkManagementClient, {
+  CustomIPPrefixesUpdateTagsParameters
+} from "@azure-rest/arm-network";
+import { DefaultAzureCredential } from "@azure/identity";
+import * as dotenv from "dotenv";
+
+dotenv.config();
+
+/**
+ * This sample demonstrates how to Updates custom IP prefix tags.
+ *
+ * @summary Updates custom IP prefix tags.
+ * x-ms-original-file: specification/network/resource-manager/Microsoft.Network/stable/2022-05-01/examples/CustomIpPrefixUpdateTags.json
+ */
+async function updatePublicIPAddressTags() {
+  const credential = new DefaultAzureCredential();
+  const client = createNetworkManagementClient(credential);
+  const subscriptionId = "";
+  const resourceGroupName = "rg1";
+  const customIpPrefixName = "test-customipprefix";
+  const options: CustomIPPrefixesUpdateTagsParameters = {
+    body: { tags: { tag1: "value1", tag2: "value2" } },
+    queryParameters: { "api-version": "2022-05-01" }
+  };
+  const result = await client
+    .path(
+      "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Network/customIpPrefixes/{customIpPrefixName}",
+      subscriptionId,
+      resourceGroupName,
+      customIpPrefixName
+    )
+    .patch(options);
+  console.log(result);
+}
+
+updatePublicIPAddressTags().catch(console.error);