+++ sdk/network/arm-network-rest/samples/v1-beta/typescript/src/vpnGatewaysStopPacketCaptureSample.ts
@@ -1,0 +1,47 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT license.
+
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+import createNetworkManagementClient, {
+  VpnGatewaysStopPacketCaptureParameters,
+  getLongRunningPoller
+} from "@azure-rest/arm-network";
+import { DefaultAzureCredential } from "@azure/identity";
+import * as dotenv from "dotenv";
+
+dotenv.config();
+
+/**
+ * This sample demonstrates how to Stops packet capture on vpn gateway in the specified resource group.
+ *
+ * @summary Stops packet capture on vpn gateway in the specified resource group.
+ * x-ms-original-file: specification/network/resource-manager/Microsoft.Network/stable/2022-05-01/examples/VpnGatewayStopPacketCapture.json
+ */
+async function stopPacketCaptureOnVpnGateway() {
+  const credential = new DefaultAzureCredential();
+  const client = createNetworkManagementClient(credential);
+  const subscriptionId = "";
+  const resourceGroupName = "rg1";
+  const gatewayName = "vpngw";
+  const options: VpnGatewaysStopPacketCaptureParameters = {
+    body: {
+      sasUrl:
+        "https://teststorage.blob.core.windows.net/?sv=2018-03-28&ss=bfqt&srt=sco&sp=rwdlacup&se=2019-09-13T07:44:05Z&st=2019-09-06T23:44:05Z&spr=https&sig=V1h9D1riltvZMI69d6ihENnFo%2FrCvTqGgjO2lf%2FVBhE%3D"
+    },
+    queryParameters: { "api-version": "2022-05-01" }
+  };
+  const initialResponse = await client
+    .path(
+      "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Network/vpnGateways/{gatewayName}/stoppacketcapture",
+      subscriptionId,
+      resourceGroupName,
+      gatewayName
+    )
+    .post(options);
+  const poller = getLongRunningPoller(client, initialResponse);
+  const result = await poller.pollUntilDone();
+  console.log(result);
+}
+
+stopPacketCaptureOnVpnGateway().catch(console.error);