+++ sdk/apimanagement/arm-apimanagement/CHANGELOG.md
@@ -1,5 +1,21 @@
 # Release History
-    
+
+## 8.1.2 (Unreleased)
+
+### Features Added
+
+### Breaking Changes
+
+### Bugs Fixed
+
+### Other Changes
+
+## 8.1.1 (2022-09-30)
+
+**Bugs Fixed**
+
+  -  fix better user experience of credential scopes in government cloud
+
 ## 8.1.0 (2022-08-03)
     
 **Features**